+++ modules/hanna-react/src/utils.ts
@@ -4,6 +4,7 @@ export * from './utils/HannaUIState.js';
 export * from './utils/useDidChange.js';
 export * from './utils/useFormatMonitor.js';
 export * from './utils/useGetSVGtext.js';
+export * from './utils/useMenuToggling.js';
 export * from './utils/useMixedControlState.js';
 export * from './utils/useScrollbarWidthCSSVar.js';